refactor: RefreshOverlay 적용 위해 루트 layout 수정 및 /customer/layout.tsx 제거
